# target/classes/swagger.yaml
swagger: "2.0"
info:
  description: "当前项目提供了获取个人信息页面需要的所有数据的多个API函数以及对应的接口(重点关注参数和返回值)"
  version: "1.0.0"
  title: "Web开发与实践"
  # termsOfService: "http://swagger.io/terms/"
  contact:
    email: "gatesma@foxmail.com"
    url: "http://gatesma.cn"
# host: "gatesma.cn"
...
schemes:
  - "http"
paths:
...
  /personal/set_basicProfile:
    post:
      tags:
        - "personal"
      summary: "设置最基本的个人信息"
      description: "名字、生日、邮箱、电话、FREELANCE"
      operationId: "basicprofileset"
      consumes:
        - "application/json"
      produces:
        - "application/json"
      parameters:
        - in: "body"
          name: "body"
          required: true
          schema:
            $ref: '#/definitions/BasicProfileSetRequest'
      responses:
        200:
          description: OK
          schema:
            $ref: '#/definitions/BasicProfileSetRet'
...
  /personal/set_scorelist:
    post:
      tags:
        - "personal"
      summary: "设置个人成绩"
      description: "设置个人成绩接口"
      operationId: "scorelistset"
      consumes:
        - "application/json"
      produces:
        - "application/json"
      parameters:
        - in: "body"
          name: "body"
          required: true
          schema:
            $ref: '#/definitions/ScoreListSetRequest'
      responses:
        200:
          description: OK
          schema:
            $ref: '#/definitions/ScoreListSetRet'
...
  /personal/set_projectlist:
    post:
      tags:
        - "personal"
      summary: "设置个人经历项目"
      description: "设置个人经历项目接口"
      operationId: "projectlistset"
      consumes:
        - "application/json"
      produces:
        - "application/json"
      parameters:
        - in: "body"
          name: "body"
          required: true
          schema:
            $ref: '#/definitions/ProjectListSetRequest'
      responses:
        200:
          description: OK
          schema:
            $ref: '#/definitions/ProjectListSetRet'
...
definitions:
  ReturnCode:
    type: object
    required:
      - code
      - message
    properties:
      code:
        type: integer
        format: int32
        example: 0
      message:
        type: string
        example: OK

  # Request
...
  BasicProfileSetRequest:
    properties:
      name:
        type: string
        description: "个人名字"
      born_time:
        type: string
        description: "出生时间，格式：yyyy-MM-dd HH:mm:ss"
      email:
        type: string
        description: "邮箱"
      phone:
        type: string
        description: "电话"
      freelance:
        type: string
        description: "现在是否自由"
  ScoreListSetRequest:
    required:
      - list
    properties:
      list:
        type: array
        items:
          $ref: '#/definitions/score'
  ProjectListSetRequest:
    required:
      - list
    properties:
      list:
        type: array
        items:
          $ref: '#/definitions/project'
...
  score:
    properties:
      name:
        type: string
        description: "科目名称"
      score:
        type: integer
        minimum: 0
        maximum: 100
        example: 90
        description: "成绩"
  project:
    properties:
      name:
        type: string
        description: "项目名称"
      time:
        type: string
        description: "项目时间"
      content:
        type: string
        description: "项目内容"
...
  ScoreListSetRet:
    allOf:
      - $ref: '#/definitions/ReturnCode'
...
  BasicProfileSetRet:
    allOf:
      - $ref: '#/definitions/ReturnCode'
  ProjectListSetRet:
    allOf:
      - $ref: '#/definitions/ReturnCode'
...